hack to get last of the np2 tests passing
revisit this one

+++ src/aspire/reconstruction/mean.py
@@ -139,7 +139,7 @@ class WeightedVolumesEstimator(Estimator):
                     if j != k:
                         kernel[j, k] += batch_kernel
 
-        kermat_f = np.zeros((self.r, self.r, _2L, _2L, _2L))
+        kermat_f = np.zeros((self.r, self.r, _2L, _2L, _2L), dtype=self.dtype)
         logger.info("Computing non-centered Fourier Transform Kernel Mat")
         for k in range(self.r):
             for j in range(self.r):
@@ -210,6 +210,7 @@ class WeightedVolumesEstimator(Estimator):
             precond_kernel = self.precond_kernel
             if regularizer > 0:
                 precond_kernel += regularizer
+
             M = LinearOperator(
                 (self.r * count, self.r * count),
                 matvec=partial(self.apply_kernel, kernel=precond_kernel),
@@ -262,7 +263,7 @@ class WeightedVolumesEstimator(Estimator):
                 f"Conjugate gradient unable to converge after {info} iterations."
             )
 
-        return x.reshape(self.r, self.basis.count)
+        return x.reshape(self.r, self.basis.count).astype(self.dtype, copy=None)
 
     def apply_kernel(self, vol_coef, kernel=None):
         """
@@ -283,8 +284,7 @@ class WeightedVolumesEstimator(Estimator):
         vols_out = Volume(
             np.zeros((self.r, self.src.L, self.src.L, self.src.L), dtype=self.dtype)
         )
-
-        vol = Coef(self.basis, vol_coef).evaluate()
+        vol = Coef(self.basis, vol_coef.astype(self.dtype, copy=False)).evaluate()
 
         for k in range(self.r):
             for j in range(self.r):
